*Added several basic gui classes like drawing box/circle only with the GX engine without the need of a resource *Created a checkbox class with them *Changed the prompt window to a class *Changed cover download to a checkbox selection of what you want to download. Changed discart download option to only switch between what is prioritized (original/customs) *Fixed a little alignment issue of the progressbar *Reverted the change for check disc in drive. Some drives couldn't handle it in IOS58 mode :\. *Changed playcount from an 8 bit to a 32 bit counter (got 8bit by mistake) which limited it to 255 *a few little optimisations

+++ source/settings/CGameStatistics.h
@@ -11,7 +11,7 @@ typedef struct _Stats
 {
     char id[7];
     u8 FavoriteRank;
-    u8 PlayCount;
+    int PlayCount;
 } GameStatus;
 
 class CGameStatistics
